pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grafana-stack-templates"
version = "0.1.0"
description = "Curated catalog of ready-to-apply Grafana Cloud building blocks: dashboards, alert rules, SM checks, Slack templates."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Tarun Saini"}]
keywords = ["grafana", "monitoring", "alerting", "synthetic-monitoring", "slack"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: System :: Monitoring",
]
dependencies = [
  "requests>=2.28",
  "pyyaml>=6.0",
  "click>=8.0",
  "jinja2>=3.1",
]

[project.scripts]
gst = "grafana_stack_templates.cli:main"

[project.urls]
Homepage = "https://github.com/Tarunrj99/grafana-stack-templates"
Issues = "https://github.com/Tarunrj99/grafana-stack-templates/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/grafana_stack_templates"]
